requirements.txt
grpcio==1.54.2
langchain==0.0.190
protobuf==4.23.2
